requirements.txt
flask
google-cloud-storage
google-cloud-secret-manager
google-auth
google-auth-oauthlib